Restructure project: move app files to src/app directory and add Docker support
- Move app module, controller, service, and tests from src/ to src/app/
- Update imports in main.ts and e2e tests to reflect new structure
- Add Docker configuration files (Dockerfile, docker-compose.yml)
- Add .env.example template and ignore .env.testing
- Remove default NestJS boilerplate files from root src directory
