+++ .env.example
@@ -1,0 +1,9 @@
+NODE_ENV=development
+APP_PORT=9999
+DB_CONNECTION=postgres
+DB_HOST=db
+DB_PORT=5432
+FORWARD_DB_PORT=33484
+DB_DATABASE=good_morning
+DB_USERNAME=user
+DB_PASSWORD=password
