+++ .gitignore
@@ -37,6 +37,7 @@ lerna-debug.log*
 
 # dotenv environment variable files
 .env
+.env.testing
 .env.development.local
 .env.test.local
 .env.production.local
