+++ docker-compose.yml
@@ -1,0 +1,100 @@
+services:
+  app:
+    build: .
+    restart: always
+    ports:
+      - '${APP_PORT}:${APP_PORT}'
+    platform: linux/amd64
+    volumes:
+      - '.:/home/app'
+    command: bash -c "rm -rf node_modules dist && npm install && npm run start:dev"
+    depends_on:
+      db:
+        condition: service_healthy
+      redis:
+        condition: service_healthy
+    networks:
+      - app-net
+  db:
+    image: 'mysql/mysql-server:8.0'
+    ports:
+      - '${FORWARD_DB_PORT:-3306}:3306'
+    environment:
+      MYSQL_ROOT_PASSWORD: '${DB_PASSWORD}'
+      MYSQL_ROOT_HOST: '%'
+      MYSQL_DATABASE: '${DB_DATABASE}'
+      MYSQL_USER: '${DB_USERNAME}'
+      MYSQL_PASSWORD: '${DB_PASSWORD}'
+      MYSQL_ALLOW_EMPTY_PASSWORD: 1
+    volumes:
+      - 'app-mysql:/var/lib/mysql'
+    networks:
+      - app-net
+    healthcheck:
+      test: ['CMD', 'mysqladmin', 'ping', '-p${DB_PASSWORD}']
+      retries: 10
+      interval: 5s
+      timeout: 5s
+      start_period: 5s
+  db-test:
+    image: 'mysql/mysql-server:8.0'
+    environment:
+      MYSQL_ROOT_PASSWORD: '${DB_PASSWORD}'
+      MYSQL_ROOT_HOST: '%'
+      MYSQL_DATABASE: '${DB_DATABASE}'
+      MYSQL_USER: '${DB_USERNAME}'
+      MYSQL_PASSWORD: '${DB_PASSWORD}'
+      MYSQL_ALLOW_EMPTY_PASSWORD: 1
+    volumes:
+      - 'app-mysql-testing:/var/lib/mysql'
+    networks:
+      - app-net
+    healthcheck:
+      test: ['CMD', 'mysqladmin', 'ping', '-p${DB_PASSWORD}']
+      retries: 10
+      interval: 5s
+      timeout: 5s
+      start_period: 5s
+  redis:
+    image: 'redis:alpine'
+    ports:
+      - '${FORWARD_REDIS_PORT:-6379}:6379'
+    volumes:
+      - 'app-redis:/data'
+    networks:
+      - app-net
+    healthcheck:
+      test:
+        - CMD
+        - redis-cli
+        - ping
+      retries: 3
+      timeout: 5s
+  redis-test:
+    image: 'redis:alpine'
+    # ports:
+    #   - '${FORWARD_REDIS_PORT:-6379}:6379'
+    volumes:
+      - 'app-redis-testing:/data'
+    networks:
+      - app-net
+    healthcheck:
+      test:
+        - CMD
+        - redis-cli
+        - ping
+      retries: 3
+      timeout: 5s
+
+networks:
+  app-net:
+    driver: bridge
+volumes:
+  app-mysql:
+    driver: local
+  app-mysql-testing:
+    driver: local
+  app-redis:
+    driver: local
+  app-redis-testing:
+    driver: local
